backend/tests/load/requirements.txt
locust==2.19.1
pytest==7.4.3
python-dotenv==1.0.0
requests==2.31.0
matplotlib==3.8.2
pandas==2.1.4
